+++ src/components/gl-wrapper/__tests__/gl-wrapper.test.js
@@ -1,5 +1,5 @@
 import renderer from 'react-test-renderer';
-import { testCases } from './gl-wrapper-test-cases.js';
+import { testCases, noRenderCases } from './gl-wrapper-test-cases.js';
 
 describe('gl-wrapper', () => {
   describe(testCases.basic.description, () => {
@@ -9,6 +9,38 @@ describe('gl-wrapper', () => {
 
     beforeEach(() => {
       testCase = testCases.basic;
+      wrapper = renderer.create(testCase.element);
+      tree = wrapper.toJSON();
+    });
+
+    test('renders as expected', () => {
+      expect(tree).toMatchSnapshot();
+    });
+  });
+
+  describe(noRenderCases.ie.description, () => {
+    let testCase;
+    let wrapper;
+    let tree;
+
+    beforeEach(() => {
+      testCase = noRenderCases.ie;
+      wrapper = renderer.create(testCase.element);
+      tree = wrapper.toJSON();
+    });
+
+    test('renders as expected', () => {
+      expect(tree).toMatchSnapshot();
+    });
+  });
+
+  describe(noRenderCases.webgl.description, () => {
+    let testCase;
+    let wrapper;
+    let tree;
+
+    beforeEach(() => {
+      testCase = noRenderCases.webgl;
       wrapper = renderer.create(testCase.element);
       tree = wrapper.toJSON();
     });